refactor(relay): extract relay-client and relay-server packages
Split relay transport client from @pleaseai/agent-core into
@pleaseai/relay-client and relay server from apps/relay-worker
into @pleaseai/relay-server for independent reuse.

- Move RelayTransport, RelayConfig, RelayEnvelope to relay-client
- Move RelayParty, Env, signature verification to relay-server
- Core re-exports relay-client types for backward compatibility
- relay-worker imports from relay-server package
