+++ packages/core/src/orchestrator.ts
@@ -14,7 +14,7 @@ import { toDispatchLockKey } from './dispatch-lock'
 import { createLabelService } from './label'
 import { createLogger } from './logger'
 import { buildContinuationPrompt, buildPrompt, isPromptBuildError } from './prompt-builder'
-import { RelayTransport } from './relay-transport'
+import { RelayTransport } from '@pleaseai/relay-client'
 import { createTrackerAdapter, formatTrackerError, isTrackerError } from './tracker/index'
 import { isWorkflowError, loadWorkflow } from './workflow'
 import { configureRemoteAuth, createWorkspace, removeRemoteAuth, removeWorkspace, runAfterRunHook, runBeforeRunHook } from './workspace'
