+++ packages/relay-client/src/index.ts
@@ -1,0 +1,2 @@
+export { RelayTransport } from './relay-transport'
+export type { RelayConfig, RelayEnvelope } from './types'
