+++ packages/relay-client/src/types.ts
@@ -1,0 +1,13 @@
+export interface RelayConfig {
+  url: string | null
+  token: string | null
+  room: string | null
+  secret: string | null
+}
+
+export interface RelayEnvelope {
+  type: string
+  event_id?: string
+  event?: string
+  action?: string | null
+}
