+++ packages/relay-client/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "types": ["bun-types"],
+    "outDir": "./dist"
+  },
+  "include": ["src"]
+}
